2024/b01lers CTF 2024/web/imagehost/requirements.txt
Pillow==10.3.0
cryptography==42.0.5
pyjwt==2.8.0
